SageMath version 10.0, Release Date: 2023-05-20

* tag '10.0': (26233 commits)
  Updated SageMath version to 10.0
  Remove tab characters
  Updated SageMath version to 10.0.rc3
  RAM for WSL in  README.md
  build/pkgs/python3/spkg-configure.m4: Add depcheck for zlib
  remark that WSL needs a lot of RAM
  A typo
  Replace git to Git
  Remove defunct references
  Minor edits of index.rst
  Minor edits on workspace.rst
  Edit git fetch -b ...
  Fix upstream
  Also mention GCM
  Remove a reference
  Replace shell-session with console to test
  Remove duplicates in git_setup
  Some moving around
  Change master to develop
  Edit conflict resolution section
  ...

+++ src/sage_setup/clean.py
@@ -2,7 +2,10 @@
 Clean the Install Dir
 """
 #*****************************************************************************
-#       Copyright (C) 2014 Volker Braun <vbraun.name@gmail.com>
+#       Copyright (C) 2014      Volker Braun <vbraun.name@gmail.com>
+#                     2015      Jeroen Demeyer
+#                     2017      Erik M. Bray
+#                     2020-2022 Matthias Koeppe
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -13,13 +16,10 @@ Clean the Install Dir
 
 
 import os
-import six
+import importlib.util
 
-if not six.PY2:
-    import importlib.util
-
+from sage.misc.package_dir import SourceDistributionFilter
 from sage_setup.find import installed_files_by_module, get_extensions
-
 
 
 def _remove(file_set, module_base, to_remove):
@@ -63,13 +63,11 @@ def _remove(file_set, module_base, to_remove):
 
         remove = [filename]
 
-        if not six.PY2:
-            remove.append(importlib.util.cache_from_source(filename))
-
+        remove.append(importlib.util.cache_from_source(filename))
         file_set.difference_update(remove)
 
 
-def _find_stale_files(site_packages, python_packages, python_modules, ext_modules, data_files):
+def _find_stale_files(site_packages, python_packages, python_modules, ext_modules, data_files, nobase_data_files=()):
     """
     Find stale files
 
@@ -82,23 +80,30 @@ def _find_stale_files(site_packages, python_packages, python_modules, ext_module
     course. We check that when the doctest is being run, that is,
     after installation, there are no stale files::
 
-        sage: from sage.env import SAGE_SRC, SAGE_LIB
-        sage: cythonized_dir = os.path.join(SAGE_SRC, "build", "cythonized")
+        sage: from sage.env import SAGE_SRC, SAGE_LIB, SAGE_ROOT
+        sage: from sage_setup.find import _cythonized_dir
+        sage: cythonized_dir = _cythonized_dir(SAGE_SRC)
         sage: from sage_setup.find import find_python_sources, find_extra_files
-        sage: python_packages, python_modules = find_python_sources(
+        sage: python_packages, python_modules, cython_modules = find_python_sources(
         ....:     SAGE_SRC, ['sage', 'sage_setup'])
-        sage: extra_files = list(find_extra_files(SAGE_SRC,
-        ....:     ['sage', 'sage_setup'], cythonized_dir, []).items())
+        sage: extra_files = find_extra_files(SAGE_SRC,
+        ....:     ['sage', 'sage_setup'], cythonized_dir, [])
+        sage: from importlib.metadata import files
+        sage: for f in files('sagemath-standard'):
+        ....:     dir = os.path.dirname(str(f))
+        ....:     extra_files[dir] = extra_files.get(dir, [])
+        ....:     extra_files[dir].append(str(f))
+        sage: extra_files = list(extra_files.items())
         sage: from sage_setup.clean import _find_stale_files
 
-    TODO: move ``module_list.py`` into ``sage_setup`` and also check
-    extension modules::
+    TODO: Also check extension modules::
 
         sage: stale_iter = _find_stale_files(SAGE_LIB, python_packages, python_modules, [], extra_files)
-        sage: from sage.misc.sageinspect import loadable_module_extension
-        sage: skip_extensions = (loadable_module_extension(),)
+        sage: from importlib.machinery import EXTENSION_SUFFIXES
+        sage: skip_extensions = tuple(EXTENSION_SUFFIXES)
         sage: for f in stale_iter:
         ....:     if f.endswith(skip_extensions): continue
+        ....:     if '/ext_data/' in f: continue
         ....:     print('Found stale file: ' + f)
     """
 
@@ -106,7 +111,7 @@ def _find_stale_files(site_packages, python_packages, python_modules, ext_module
     CEXTMOD_EXTS = get_extensions('extension')
     INIT_FILES = tuple('__init__' + x for x in PYMOD_EXTS)
 
-    module_files = installed_files_by_module(site_packages, ['sage', 'sage_setup'])
+    module_files = installed_files_by_module(site_packages, ['sage'])
 
     for mod in python_packages:
         try:
@@ -134,6 +139,9 @@ def _find_stale_files(site_packages, python_packages, python_modules, ext_module
     for dir, files in data_files:
         for f in files:
             installed_files.add(os.path.join(dir, os.path.basename(f)))
+    for dir, files in nobase_data_files:
+        for f in files:
+            installed_files.add(f)
 
     for files in module_files.values():
         for f in files:
@@ -141,7 +149,8 @@ def _find_stale_files(site_packages, python_packages, python_modules, ext_module
                 yield f
 
 
-def clean_install_dir(site_packages, python_packages, python_modules, ext_modules, data_files):
+def clean_install_dir(site_packages, python_packages, python_modules, ext_modules, data_files, nobase_data_files, *,
+                      distributions=None, exclude_distributions=None):
     """
     Delete all modules that are **not** being installed
 
@@ -166,11 +175,23 @@ def clean_install_dir(site_packages, python_packages, python_modules, ext_module
       output of ``cythonize``.
 
     - ``data_files`` -- a list of (installation directory, files) pairs,
-      like the ``data_files`` argument to distutils' ``setup()``.
+      like the ``data_files`` argument to distutils' ``setup()``. Only
+      the basename of the files is used.
+
+    - ``nobase_data_files`` -- a list of (installation directory, files)
+      pairs. The files are expected to be in a subdirectory of the
+      installation directory; the filenames are used as is.
+
+    - ``distributions`` -- (default: ``None``) if not ``None``,
+      should be a sequence or set of strings: only clean files whose
+      ``distribution`` (from a ``# sage_setup: distribution = PACKAGE``
+      directive in the file) is an element of ``distributions``.
     """
+    distribution_filter = SourceDistributionFilter(distributions, exclude_distributions)
     stale_file_iter = _find_stale_files(
-        site_packages, python_packages, python_modules, ext_modules, data_files)
+        site_packages, python_packages, python_modules, ext_modules, data_files, nobase_data_files)
     for f in stale_file_iter:
         f = os.path.join(site_packages, f)
-        print('Cleaning up stale file: {0}'.format(f))
-        os.unlink(f)
+        if f in distribution_filter:
+            print('Cleaning up stale file: {0}'.format(f))
+            os.unlink(f)
